+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # FSL installer script release history
 
 
+# 3.13.0 (Wednesday 19th June 2024)
+
+ - The `fslinstaller.py` script will now instruct `conda` to install versions
+   of CUDA packages which are compatible with a local GPU, or which are
+   compatible with a CUDA version specified by the new `--cuda X.Y`
+   command-line option.
+
+
 # 3.12.2 (Friday 14th June 2024)
 
  - Check for more network errors that may be able to be solved by retrying the